work on player data

+++ bukkit/src/main/kotlin/dev/jaims/moducore/bukkit/api/manager/storage/FileStorageManager.kt
@@ -44,7 +44,7 @@ class FileStorageManager(private val plugin: ModuCore) : StorageManager() {
     override var updateTask = plugin.launchAsync {
         saveAllData(playerDataCache)
         while (true) {
-            delay(60 * 1000)
+            delay((60 * 1000).toLong())
         }
     }
